Rewrite markdown implementation
We are now using parss to parse markdown, with proper error handling.
The syntax is also simplified a little bit. It's also more "strongly
typed", i.e. there are more errors.

Note that errors are "reported" in the compiled document. This is kind
of great - you run the server, open your document in the browser, edit,
then see the errors directly in the output.

Importantly, we've replaced some partial functions (in Compile.hs) with
this error reporting. It's important that the server never crashes!

We're also much more explicit about titles and dates, which are needed
for indexing.

I've also removed the weird markdown header/footer and just used HTML.
The idea is to only use markdown for articles, but HTML for everything
else. Writing HTML is not a problem unless it's for a full article.