Be more responsive
Not sure if this is 100% correct/reliable/accessible/portable but it
seems like a really nice solution because it's responsive based on
character width.

Not sure why there's a ~20ch difference between the viewport width and
body width. The margins are only maybe 5 characters in total.

Also simplified some stuff.